хук useLike - общая логика лайков  в кастомхуке
