+++ src/components/bar/TrackPlayLikeDiz/trackplaylikediz.jsx
@@ -3,24 +3,28 @@ import * as S from './trackplaylikediz.style'
 import { useSelector } from 'react-redux'
 import {
     useGetAllFavoritesQuery,
-    useAddFavoriteTrackMutation,
-    useDeleteFavoriteTrackMutation,
+    // useAddFavoriteTrackMutation,
+    // useDeleteFavoriteTrackMutation,
 } from '../../../store/services/favorite'
+import { useLike } from '../../../hooks/useLike'
 
 function TrackPlayLikeDiz() {
     const token = useSelector((s) => s.state.accessToken)
     const currentTrack = useSelector((s) => s.state.currentTrack)
     const { data: favoritesPlaylist } = useGetAllFavoritesQuery(token)
-    const [addFavorite] = useAddFavoriteTrackMutation()
-    const [deleteFavorite] = useDeleteFavoriteTrackMutation()
+    // const [addFavorite] = useAddFavoriteTrackMutation()
+    // const [deleteFavorite] = useDeleteFavoriteTrackMutation()
+    const [handleClickToLike,handleClickToDizLike] = useLike()
 
     const favoriteArrayByID = favoritesPlaylist?.map((elem) => elem.id)
     const likedStatus = favoriteArrayByID?.includes(currentTrack.id)
 
     function handleClick() {
         likedStatus
-            ? deleteFavorite({ id: currentTrack.id, accessToken: token })
-            : addFavorite({ id: currentTrack.id, accessToken: token })
+            // ? deleteFavorite({ id: currentTrack.id, accessToken: token })
+            // : addFavorite({ id: currentTrack.id, accessToken: token })
+            ? handleClickToDizLike(currentTrack)
+            : handleClickToLike(currentTrack)  
     }
 
     return (
